// packages/app/lib/types/animation.ts
export enum ANIMATION_ACTION {
  NONE = 'none',
  IDLE = 'idle',
  SHELF_IDLE = 'shelf_idle',
  SINK_IDLE = 'sink_idle',
  TOILET_IDLE = 'toilet_idle',
  WALK = 'walk',
  STAIR_FALLBACK = 'stair_fallback',
  SITTING_IDLE = 'sitting_idle',
  LAYING_SLEEPING = 'laying_sleeping',
  CLOSED = 'closed',
  CLOSE_LEFT = 'close_left',
  CLOSE_RIGHT = 'close_right',
  OPEN_LEFT = 'open_left',
  OPEN_RIGHT = 'open_right',
  OPEN = 'open',
  CLOSE = 'close'
}